+++ lab08/Paint/Triangle.cpp
@@ -15,7 +15,7 @@ DrawCommand CTriangle::DrawShape() const
 	m_points.emplace_back((m_rect.left + m_rect.right) / 2, m_rect.top);
 	m_points.emplace_back(m_rect.right, m_rect.bottom);
 
-	return bind(&CDC::Polygon, placeholders::_1, m_points.data(), m_points.size());
+	return bind(&CDC::Polygon, placeholders::_1, m_points.data(), boost::numeric_cast<int>(m_points.size()));
 }
 
 bool CTriangle::IsPointInShape(CPoint const & point) const
@@ -26,7 +26,7 @@ bool CTriangle::IsPointInShape(CPoint const & point) const
 	m_points.emplace_back(m_rect.right, m_rect.bottom);
 	
 	CRgn rgn;
-	rgn.CreatePolygonRgn(m_points.data(), m_points.size(), WINDING);
+	rgn.CreatePolygonRgn(m_points.data(), boost::numeric_cast<int>(m_points.size()), WINDING);
 
 	return rgn.PtInRegion(point) != 0;
 }
